% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{a}
\alias{a}
\title{Quality control module
Functionality to ensure data issues are exposed
# Check the light intensity is stable through the experiment
 An unstable IR strip can cause the data to be corrupted because the ethoscope
 tracker is designed to track pixel intensity changes.
 In normal conditions, they follow the animal behavior,
 however an unstable IR strip can produce intensity changes independent of the animal
 which confuses the tracker and bogs the results}
\usage{
a()
}
\description{
## Check the flies move through out the tube during SD,
as opposed to stay in only a tiny fraction of the tube
SD is another component contributing to background instability for obvious reasons:
if the tube moves, there will be several pixels in the ROI that will change,
independent of the animal. The tracker is robust,
but sometimes this changes are noisy enough to again confuse the tracker.
This, combined with the fact that a very sleepy fly can actually not react
to a moving tube, drive the labeling algorithm to label a fly as awake when it really is asleep
}
