% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{apply_pareto_rule}
\alias{apply_pareto_rule}
\title{Wrap bin_apply_all and pareto_sd}
\usage{
velocity_correction_coef <- 0.0048 # change as needed
dt_raw <- load_ethoscope(linked_metadata)
dt <- sleepr::sleep_annotation(dt, velocity_correction_coef=velocity_correction_coef)
dt_binned <- behavr::bin_apply_all(data = dt, y = "asleep", x = "t", x_bin_length = behavr::mins(30), FUN=mean)
dt_binned <- apply_pareto_rule(dt, dt_binned)
ggplot(dt_binned, aes(x=t, y=asleep)) + stat_pop_etho()#'
}
\arguments{
\item{scored_dataset}{behavr timeseries data with at least columns x and t}

\item{binned_dataset}{behavr timeseries produced by applying bin_apply_all with a function other than pareto_sd,
typically the mean}

\item{sd_only}{If TRUE, this refinement is only applicable during SD}
}
\value{
A new binned_dataset with an extra field 'pareto' stating whether the animal fulfills the pareto principle of
awakeness for the given bin.
}
\description{
Wrap bin_apply_all and pareto_sd
}
\seealso{
Other pareto_sd: 
\code{\link{pareto_sd}()}
}
\concept{pareto_sd}
