% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{as_character_column}
\alias{as_character_column}
\title{Coerce a column to character}
\usage{
as_character_column(metadata, column_name)
}
\description{
Make some columns that could be considered numeric character
because that way thei handling is easier in esquisse
}
