% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify.R
\name{behavr_placeholder}
\alias{behavr_placeholder}
\title{Initialize a sensible placeholder with the right fields
but no data}
\usage{
behavr_placeholder()
}
\description{
Initialize a sensible placeholder with the right fields
but no data
}
