% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_data}
\alias{bin_data}
\title{Optionally bin data}
\usage{
bin_data(data, do = TRUE, ...)
}
\description{
`bin_data` will run `fslggetho::ggetho_preprocess` on a provided behavr table
only if do
Even if not do, it returns default values for plot parameters that are needed by ggetho_plot
}
