% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bout_analysis}
\alias{bout_analysis}
\title{A version of sleepr::bout_analysis that allows programmatic input}
\usage{
bout_analysis(data, ...)
}
\arguments{
\item{data}{\link{data.table} containing behavioural variable from or one multiple animals.
When it has a key, unique values, are assumed to represent unique individuals (e.g. in a \link{behavr} table).
Otherwise, it analysis the data as coming from a single animal. \code{data} must have a column \code{t} representing time.}
}
\description{
A version of sleepr::bout_analysis that allows programmatic input
}
\seealso{
sleepr::bout_analysis
}
