% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{datetime_filename}
\alias{datetime_filename}
\title{Prepend a filename (or any character) with a datetime signature}
\usage{
datetime_filename(filename)
}
\description{
`datetime_filename` reinforces good documentation practices across
fslretho to help users keep track of the files generated with the application
}
