% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{defineSingleGroupServer}
\alias{defineSingleGroupServer}
\title{Process user entered groups of animals}
\usage{
defineSingleGroupServer(id, rv)
}
\description{
Return a character defining an animal group
implemented via row filtering in data.tables
}
