% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-video-utils.R
\name{ethoscope_imager}
\alias{ethoscope_imager}
\title{Save ethoscope snapshots to disk with Python
The sqlite3 library in Python is well suited to interface
with dbfiles produced in the ethoscope platform}
\usage{
ethoscope_imager(path, id = NULL, video = F, fps = NULL)
}
\arguments{
\item{path}{Absolute path to a sqlite3 file produced with an ethoscope}

\item{id}{Numeric vector of ids matching ids in the IMG_SNAPSHOTS table of the dbfile}

\item{video}{If TRUE, produce a video with the supplied ids.
Alternatively, if no ids are passed,
then with all the cached (already saved outside of the dbfile) snapshots}

\item{fps}{Frame rate of the resulting video}
}
\description{
Save ethoscope snapshots to disk with Python
The sqlite3 library in Python is well suited to interface
with dbfiles produced in the ethoscope platform
}
