% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify.R
\name{extend_dataset}
\alias{extend_dataset}
\title{Populate a pair of placeholder data.tables with the loaded data}
\usage{
extend_dataset(dt, metadata, new_dataset, monitor)
}
\description{
Populate a dependency generating pair of placeholders
with the data loaded by the user.
The placeholders contain the standard ethoscope and dam fields
and are extended with more if the provided data carries them
}
