% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify.R
\name{extend_datatable}
\alias{extend_datatable}
\title{Add  fields available in a table to another table missing  them}
\usage{
extend_datatable(x, xx)
}
\arguments{
\item{x}{Receiving table}

\item{xx}{Donating table}
}
\description{
Add  fields available in a table to another table missing  them
}
