% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fortify}
\alias{fortify}
\title{Make a writable behavr object}
\usage{
fortify(data, meta = FALSE)
}
\arguments{
\item{data}{A rejoined behavr or any data.table object}

\item{meta}{If TRUE, repeat also for the metadata
TODO Possibly this should be part of a fwrite_behavr function in behavr}
}
\description{
Make a writable behavr object
}
