% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{FSLRethoConfiguration}
\alias{FSLRethoConfiguration}
\title{Configuration of FSLRetho}
\description{
Load and update the configuration of the program
}
\details{
Anywhere in FSLRetho `FSLRethoConfiguration` can be instantiated so
default values for hardware defined parameters can be set dynamically
via a config file
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{content}}{List of configuration items}

\item{\code{config_file}}{Default configuration file path
Initialize a configuration instance
Called by running `FSLRethoConfiguration$new()`}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FSLRethoConfiguration$new()}}
\item \href{#method-save}{\code{FSLRethoConfiguration$save()}}
\item \href{#method-load}{\code{FSLRethoConfiguration$load()}}
\item \href{#method-clone}{\code{FSLRethoConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSLRethoConfiguration$new(config_file = "/etc/fslretho.conf")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_file}}{Path to the default configuration file
Save the configuration in to a config_file

Save the configuration stored in self$content
If the passed config_file is null, use the instance's default}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSLRethoConfiguration$save(config_file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_file}}{Configuration file path
Load a a configuration from a config_file
If the passed config_file is null, use the instance's default
If the passed file does not exist create it with the conf in the default file
If it exists, load its contents and update the configuration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\if{latex}{\out{\hypertarget{method-load}{}}}
\subsection{Method \code{load()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSLRethoConfiguration$load(config_file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_file}}{Configuration file path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSLRethoConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
