% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\docType{data}
\name{FUN_choices}
\alias{FUN_choices}
\title{Bin time series of boolean variables of high frequency (>= 1/min)
so they become a numerical time series of a lower frequency
The meaning of the new time series depends of the binning function used
i.e. if mean is used, the time series is the fraction of time spent in the
TRUE state in each window
Relevant for moving, asleep and interactions variables}
\format{
An object of class \code{character} of length 5.
}
\usage{
FUN_choices
}
\description{
Bin time series of boolean variables of high frequency (>= 1/min)
so they become a numerical time series of a lower frequency
The meaning of the new time series depends of the binning function used
i.e. if mean is used, the time series is the fraction of time spent in the
TRUE state in each window
Relevant for moving, asleep and interactions variables
}
\keyword{datasets}
