% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_monitor_name}
\alias{get_monitor_name}
\title{Given a metadata file, infer what type of monitor it is}
\usage{
get_monitor_name(metadata_path)
}
\arguments{
\item{metadata_path}{Path to a metadata file}
}
\value{
character either dam or ethoscope
}
\description{
Given a metadata file, infer what type of monitor it is
}
