% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{get_progress_bar}
\alias{get_progress_bar}
\title{Produce a Shiny progress bar}
\usage{
get_progress_bar(steps)
}
\arguments{
\item{steps}{Number of steps in the progress bar i.e. the number of times
the progress bar needs to be advanced by 1 to have it completed}
}
\description{
Produce a Shiny progress bar
}
