% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{group_data}
\alias{group_data}
\title{Assign groups to flies}
\usage{
group_data(data, groups)
}
\arguments{
\item{data}{A shiny reactiveValue with a single reactive that yields a behavr table under key input}

\item{groups}{A shiny reactiveValues object where every reactive returns a named character
than can be used to filter rows in the behavr object.}
}
\value{
A reactive that returns the same behavr table but its metadata will have
in the column "group", where the filter evals to TRUE, the value given by the name of the filter i.e.
the name bound to the character vector in each group.
}
\description{
`group_data` populates the group column of the metadata in a behavr table
so plots and downstream analyses can segregate with user provided groups
}
