% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethoscope.R
\name{load_ethoscope}
\alias{load_ethoscope}
\title{Wrapper around the most common workflow to laod ethoscope data into R using Rethomics}
\usage{
load_ethoscope(
  metadata,
  result_dir,
  reference_hour = NULL,
  updateProgress = NULL
)
}
\arguments{
\item{metadata}{A valid data.table object describing an experimental design}

\item{result_dir}{Absolute path of an ethoscope sqlite3 database}

\item{reference_hour}{hour of the day when the light turns on in the experiment
e.g. if the lights turn on at 8 AM GMT, then reference_hour = 8
If NULL, the reference_hour is parsed from the column with the same name in the metadata
If such a column is not available, then assume ZT is the same as the experiment start}

\item{updateProgress}{Optional, a function that updates an object of class shiny::Progress}
}
\value{
A raw behavr table i.e. without annotation
}
\description{
Wrapper around the most common workflow to laod ethoscope data into R using Rethomics
}
