% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Start an FSLRetho instance}
\usage{
main(display_mode = FALSE, port = NULL)
}
\arguments{
\item{display_mode}{Whether to show the code alongside the application or not}

\item{port}{Port on which fslretho2 should run}
}
\description{
Start an FSLRetho instance
}
