% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_sqlite.R
\name{make_toy_ethoscope_sqlite}
\alias{make_toy_ethoscope_sqlite}
\title{A function to create a sqlite3 file under the extdata of the package}
\usage{
make_toy_ethoscope_sqlite(nrows = 10000)
}
\description{
A function to create a sqlite3 file under the extdata of the package
}
