% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{pareto_sd}
\alias{pareto_sd}
\title{Apply Pareto rule to wake behavior during SD}
\usage{
dt_raw <- load_ethoscope(linked_metadata, FUN=sleepr::sleep_annotation) # or whatever annotation function
dt <- behavr::bin_apply_all(dt_raw, x="t", y="x", x_bin_length=behavr::mins(30), FUN=pareto_sd)
}
\arguments{
\item{x}{Position of the animal along the main axis of variation, normalized from 0 to 1}

\item{n}{Number of splits in the ROI. default 10}

\item{min_max_crop}{If TRUE, x is normalized so the minimum value is 0 and the maximum is 1
This is useful if the ROI is much wider than the tube (which would mean that some sections of the tube actually cannot be explored by the fly)}

\item{space_fraction}{Minimum fraction of tube that a wake animal explores. If the animal explores less than this, it is not awake}

\item{time_fraction}{Maximum fraction of time that a wake animal is allowed to stay in the p fraction of the tube}
}
\value{
TRUE if the tube exploration is < expected with Pareto and FALSE otherwise
A result of TRUE indicates thus a problematic data
}
\description{
Require significant tube exploration to label behavior as awake.
}
\details{
Minor movements (indicating little awakeness) or spiky artifacts during real sleep
both result in a very small fraction of the tube being explored by the animal in a given interval of time.

This function checks that the animal did not spend > time_fraction in < space_fraction of the tube while awake,
If the animal breaks this check (it spends too much time in too little space) the animal is asleep.

This make sense to do during SD because the tracking quality decreases due to
background changes produced by the SD itself
}
\seealso{
Other pareto_sd: 
\code{\link{apply_pareto_rule}()}
}
\concept{pareto_sd}
