% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_sd_daterange}
\alias{parse_sd_daterange}
\title{Incorporate daterange information to a behavr table}
\usage{
parse_sd_daterange(dt)
}
\arguments{
\item{dt}{A behavr table}
}
\value{
The same behavior table where the metadaata now features columns start_sd and end_sd
which contain the number of ms since experiment start until SD start and end
}
\description{
Incorporate daterange information to a behavr table
}
