% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-video-utils.R
\name{paste_regions}
\alias{paste_regions}
\title{Paste several images together}
\usage{
paste_regions(roi_images, roi_map)
}
\arguments{
\item{roi_images}{list where every element is a magick image object
storing the image of a ROI. The element must be named ROI_X}

\item{roi_map}{data.table providing a roi_value - side mapping}
}
\value{
magick image with all ROIs pasted together
}
\description{
Paste several images together
}
