% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhythm.R
\docType{data}
\name{PERIODOGRAM_ALGORITHMS}
\alias{PERIODOGRAM_ALGORITHMS}
\title{Cyrcadian rhythm functionality in FSLRetho}
\format{
An object of class \code{list} of length 3.
}
\usage{
PERIODOGRAM_ALGORITHMS
}
\description{
Cyrcadian rhythm functionality in FSLRetho
}
\keyword{datasets}
