% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_sqlite.R
\name{populate_roi}
\alias{populate_roi}
\title{Generate data for a fake fly and save it to a SQLite table}
\usage{
populate_roi(dbfile_path, table, ...)
}
\description{
Generate data for a fake fly and save it to a SQLite table
}
