% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premade.R
\name{premadePlotsServer}
\alias{premadePlotsServer}
\title{Premade plots without any user tweaking}
\usage{
premadePlotsServer(id, sleep_rv, interactions_rv)
}
\arguments{
\item{input_rv}{reactiveValues with a data slot containing an ethoscope behavr table}
}
\description{
Premade plots without any user tweaking
}
