% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_dam_metadata}
\alias{read_dam_metadata}
\title{Load a DAM metadata .csv into R}
\usage{
read_dam_metadata(metadata_path)
}
\arguments{
\item{metadata_path}{Path to a metadata file}
}
\description{
Load a DAM metadata .csv into R
}
