% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Load an ethoscope metadata .csv into R and perform basic validation}
\usage{
read_metadata(metadata_path)
}
\arguments{
\item{metadata_path}{Path to a metadata file}
}
\description{
Load an ethoscope metadata .csv into R and perform basic validation
}
