% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_sd_daterange}
\alias{read_sd_daterange}
\title{Query a database for interactor date range start / end}
\usage{
read_sd_daterange(meta_row)
}
\arguments{
\item{meta_row}{A row of metadata}
}
\value{
Integer vector of length 2 stating the milliseconds since experiment start
until onset and end of sleep deprivation treatment in the animal captured by the metadata
}
\description{
Query a database for interactor date range start / end
}
