% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rejoin_rv}
\alias{rejoin_rv}
\title{Transform the passed rv with slots data and name
so the content of data is not a behavr table anymore
and instead a rejoin()ed data.table}
\usage{
rejoin_rv(rv)
}
\description{
Transform the passed rv with slots data and name
so the content of data is not a behavr table anymore
and instead a rejoin()ed data.table
}
