% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducibility.R
\name{reproducibilityModuleServer}
\alias{reproducibilityModuleServer}
\title{Return a script that reproduces the fslretho results
Ideally this would be done as in esquisse with metaprogramming
but this is good enoug for now}
\usage{
reproducibilityModuleServer(id, envir = rlang::env())
}
\arguments{
\item{params}{reactiveValues with parameters relevant for the analysis}
}
\description{
Return a script that reproduces the fslretho results
Ideally this would be done as in esquisse with metaprogramming
but this is good enoug for now
}
