% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_monitor}
\alias{score_monitor}
\title{Annotate behavior coming from DAM or ethoscope}
\usage{
score_monitor(raw_data, input, monitor = c("ethoscope", "dam"))
}
\description{
Annotate behavior coming from DAM or ethoscope
}
