% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{sd_inprogress_annotation}
\alias{sd_inprogress_annotation}
\title{Annotate a behavr with a SD treatment progress
A new `sd_on` column stating the SD treatment status is added to the passed behavr}
\usage{
sd_inprogress_annotation(dt)
}
\arguments{
\item{dt}{behavr table}
}
\value{
the same behavr table with a new column `sd_on`.
}
\description{
Annotate a behavr with a SD treatment progress
A new `sd_on` column stating the SD treatment status is added to the passed behavr
}
