% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premade.R
\name{sleepInteractionsServer}
\alias{sleepInteractionsServer}
\title{Premade plots without any user tweaking}
\usage{
sleepInteractionsServer(id, data)
}
\arguments{
\item{data}{reactiveValues with a data slot containing an ethoscope behavr table}
}
\description{
Premade plots without any user tweaking
}
