% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video.R
\name{snapshotViewerServer}
\alias{snapshotViewerServer}
\title{Generate mp4 videos from a list of dbfiles
Ethoscopes automate high throughput behavioral tracking
however, their performance is not perfect.
Errors in the tracker or the SD delivery can bias the results
which is why the possibility to semi supervise the process is needed
Even if not needed, it is still useful to look at the rawest form of data: the images / videos
This module takes a set of dbfiles (from the metadata of the loaded dataset)
and produces a video that users can download.
The resulting frames are saved in the experiment_folder with name IMG_SNAPSHOTS
for manual frame-by-frame inspection}
\usage{
snapshotViewerServer(
  id,
  input_rv,
  dbfile_user = reactiveVal(NULL),
  trigger = reactiveVal(0)
)
}
\description{
Generate mp4 videos from a list of dbfiles
Ethoscopes automate high throughput behavioral tracking
however, their performance is not perfect.
Errors in the tracker or the SD delivery can bias the results
which is why the possibility to semi supervise the process is needed
Even if not needed, it is still useful to look at the rawest form of data: the images / videos
This module takes a set of dbfiles (from the metadata of the loaded dataset)
and produces a video that users can download.
The resulting frames are saved in the experiment_folder with name IMG_SNAPSHOTS
for manual frame-by-frame inspection
}
