% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{summariseBehavrServer}
\alias{summariseBehavrServer}
\title{Aggregate a categorical data over a fixed period of time}
\usage{
summariseBehavrServer(id, dt, feature = "asleep")
}
\arguments{
\item{dt}{A behavr table with a categorical column to be summarised}

\item{feature}{Name of the categorical column}

\item{input}{A shiny input object}

\item{output}{A shiny output object}

\item{session}{A shiny session object}
}
\value{
A list with elements
\itemize{
 \item data The summarised behavr table
 \item mapping
}
}
\description{
Given a statistic function and a fixed window length,
`summariseBehavr` generates a new behavr table where the original
time series is aggregated with that window length and a new column
with the result of applying the statistic function to a categorical variable
This behavr table and other parameters useful for plotting the time series
are returned in a list
}
