% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify.R
\name{unifyDatasetsServer}
\alias{unifyDatasetsServer}
\title{Choose either the ethoscope or the DAM dataset}
\usage{
unifyDatasetsServer(id = "", data)
}
\arguments{
\item{data}{A reactiveValues with an ethoscope and a dam slot}
}
\value{
A reactiveValues with slots data name and time of the selected dataset
}
\description{
Choose either the ethoscope or the DAM dataset
}
