% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{viewMetadataUI}
\alias{viewMetadataUI}
\title{Display a data table output of the uploaded metadata
It can be filtered by column values and sorted}
\usage{
viewMetadataUI(id)
}
\description{
Display a data table output of the uploaded metadata
It can be filtered by column values and sorted
}
