% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{watch_input}
\alias{watch_input}
\title{Declare an observer for every data slot in the passed reactive values
and map the data and name slots to the passed dest rv upon changes in the data slot
of each passed rv}
\usage{
watch_input(rv, ...)
}
\arguments{
\item{rv}{The reactiveValues object to be updated with the contents of the reactiveValues in ...}

\item{...}{reactiveValues objects with slots data and name}
}
\description{
Declare an observer for every data slot in the passed reactive values
and map the data and name slots to the passed dest rv upon changes in the data slot
of each passed rv
}
