% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip_database}
\alias{zip_database}
\title{Export a database of multiple files into a single .zipo}
\usage{
zip_database(zip_file, files)
}
\arguments{
\item{dbfiles}{character, absolute paths to files to be shipped into a single .zip}
}
\value{
path to created zip file in a temporary folder
}
\description{
Export a database of multiple files into a single .zipo
}
