% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fortify}
\alias{fortify}
\title{Make a writable behavr object}
\usage{
fortify(data)
}
\arguments{
\item{data}{A rejoined behavr or any data.table object
TODO Possibly this should be part of a fwrite_behavr function in behavr}
}
\description{
Make a writable behavr object
}
