% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{pareto_sd}
\alias{pareto_sd}
\alias{apply_pareto_rule}
\title{Quality control module
Functionality to ensure data issues are exposed
# Check the light intensity is stable through the experiment
 An unstable IR strip can cause the data to be corrupted because the ethoscope
 tracker is designed to track pixel intensity changes.
 In normal conditions, they follow the animal behavior,
 however an unstable IR strip can produce intensity changes independent of the animal
 which confuses the tracker and bogs the results}
\usage{
dt <- behavr::bin_apply_all(dt_raw, x="t", y="x",  x_bin_length=behavr::mins(30), FUN=pareto_sd)

apply_pareto_rule(scored_dataset, binned_dataset, ...)
}
\arguments{
\item{x}{Position of the animal along the main axis of variation, normalized from 0 to 1}

\item{n}{Number of splits in the ROI. default 10}

\item{scored_dataset}{behavr timeseries data with at least columns x and t}

\item{binned_dataset}{behavr timeseries produced by applying bin_apply_all with a function other than pareto_sd,
typically the mean}
}
\value{
TRUE if the tube exploration is < expected with Pareto and FALSE otherwise
A result of TRUE indicates thus a problematic data
Apply Pareto rule to wake behavior during SD

TRUE if the tube exploration is < expected with Pareto and FALSE otherwise
A result of TRUE indicates thus a problematic data

a new binned_dataset with an extra field 'pareto' stating whether the animal fulfills the pareto principle of
awakeness for the given bin.
}
\description{
## Check the flies move through out the tube during SD,
as opposed to stay in only a tiny fraction of the tube
SD is another component contributing to background instability for obvious reasons:
if the tube moves, there will be several pixels in the ROI that will change,
independent of the animal. The tracker is robust,
but sometimes this changes are noisy enough to again confuse the tracker.
This, combined with the fact that a very sleepy fly can actually not react
to a moving tube, drive the labeling algorithm to label a fly as awake when it really is asleep
Apply Pareto rule to wake behavior during SD
Require significant tube exploration to label behavior as awake
Minor movements (indicating little awakeness) or spiky artifacts during real sleep
both result in a very small fraction of the tube being explored by the animal in a given interval of time
This function checks at least > 20% of the tube is explored,
and otherwise labels the animal as sleep
This make sense to do during SD because the tracking quality decreases due to
background changes produced by the SD itself
}
