% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{scoreDataServer}
\alias{scoreDataServer}
\title{Shiny module to automatically score a raw behavr table}
\usage{
scoreDataServer(id, dt_raw, dataset_name)
}
\description{
Provide a multi-animal reactive behavr and return the scored version
}
