% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{scoreDataServer}
\alias{scoreDataServer}
\title{Shiny module to automatically score a raw behavr table}
\usage{
scoreDataServer(id, raw_data, trigger = reactiveVal(0))
}
\arguments{
\item{id}{Module id - character}

\item{raw_data}{A shiny reactiveValues with slots data and name}
}
\description{
Provide a multi-animal reactive behavr and return the scored version
}
