% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{scoreDataServer}
\alias{scoreDataServer}
\title{Shiny module to automatically score a raw behavr table}
\usage{
scoreDataServer(id, input_rv, pb = TRUE)
}
\arguments{
\item{id}{Module id - character}

\item{input_rv}{reactiveValues with one slot per monitor}

\item{pb}{logical, if TRUE, a shiny progress bar will track the processing of each individual}
}
\description{
Provide a multi-animal reactive behavr and return the scored version
}
